+++ crates/plotinator-test-util/src/test_file_defs.rs
@@ -51,6 +51,7 @@ macro_rules! define_utf8_test_file {
 pub mod bifrost_current;
 pub mod frame_altimeters;
 pub mod frame_inclinometers;
+pub mod frame_magnetometer;
 pub mod legacy_generator;
 pub mod mbed_motor_control;
 pub mod njord_ins;