+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for `Frame Inclinometers` HDF5 files
+- Support for `Frame Magnetometer` HDF5 files
 
 ### Changed